GopiAI-CrewAI/requirements_server.txt
flask
flask-cors
python-dotenv
crewai
crewai-tools
google-genai
langchain-core